% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Expectations.R
\name{CalcWorkResp}
\alias{CalcWorkResp}
\title{Generate Working Response}
\usage{
CalcWorkResp(y, mean, cov, gamma)
}
\arguments{
\item{y}{Vector with missing elements.}

\item{mean}{Numeric mean.}

\item{cov}{Numeric covariance.}

\item{gamma}{Numeric vector of responsibilities.}
}
\value{
Numeric working response vector.
}
\description{
Generate Working Response
}
