% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{CalHar.within_cluster_disp}
\alias{CalHar.within_cluster_disp}
\title{Within Cluster Dispersion}
\usage{
CalHar.within_cluster_disp(clust, mean)
}
\arguments{
\item{clust}{Numeric data.matrix.}

\item{mean}{Cluster mean.}
}
\value{
d x d within-cluster dispersion matrix.
}
\description{
Within Cluster Dispersion
}
