% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{ChooseK.iter}
\alias{ChooseK.iter}
\title{Attempt Model Fit and Return Quality Metrics.}
\usage{
ChooseK.iter(data, k, init_means, fix_means, init_covs, init_props, maxit, eps)
}
\arguments{
\item{data}{Numeric data matrix.}

\item{k}{Number of clusters.}

\item{init_means}{Optional list of initial mean vectors.}

\item{fix_means}{Fix the means to their starting value? Must initialize.}

\item{init_covs}{Optional list of initial covariance matrices.}

\item{init_props}{Optional vector of initial cluster proportions.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}
}
\value{
Numberic vector containing the 4 cluster quality metrics. Returns
  null if the model fails to fit.
}
\description{
Attempt Model Fit and Return Quality Metrics.
}
