% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{chooseK}
\alias{chooseK}
\title{Cluster Number Selection}
\usage{
chooseK(Y, k0 = 2, k1 = NULL, B = 100, M0 = NULL, fix.means = F,
  S0 = NULL, pi0 = NULL, maxit = 10, eps = 1e-04, report = T,
  parallel = F)
}
\arguments{
\item{Y}{Numeric data matrix.}

\item{k0}{Minimum number of clusters.}

\item{k1}{Maximum number of clusters.}

\item{B}{Bootstrap replicates.}

\item{M0}{Optional list of initial mean vectors.}

\item{fix.means}{Fix the means to their starting value? Must provide initial
values.}

\item{S0}{Optional list of initial covariance matrices.}

\item{pi0}{Optional vector of initial cluster proportions.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report bootstrap progress?}

\item{parallel}{Run in parallel? Only used if \eqn{k>1}. Must register
parallel backend first.}
}
\value{
List containing \code{Choices}, the recommended number of clusters
  according to each quality metric, and \code{Results}, the mean and standard
  error of the quality metrics at each cluster number evaluated.
}
\description{
Function to choose the number of clusters k. Examines cluster numbers between
k0 and k1. For each cluster number, generates B bootstrap data sets, fits the
Gaussian Mixture Model (\code{\link{fit.GMM}}), and calculates quality
metrics (\code{\link{clustQual}}). For each metric, determines the optimal
cluster number \code{kopt}, and the \code{k1se}, the smallest cluster number
 whose quality is within 1 SE of the optimum.
}
\examples{
\dontrun{
set.seed(100);
M = list(c(2,2),c(2,-2),c(-2,2),c(-2,-2));
Y = rGMM(n=500,d=2,k=4,M=M);
K = chooseK(Y=Y,k0=2,k1=6,B=10,maxit=10,eps=1e-4,report=T);
K$Choices;
}
}
\seealso{
See \code{\link{clustQual}} for evaluating cluster quality, and \code{\link{fit.GMM}}
for estimating the GMM with a specified cluster number.
}
