% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{ChooseK.recommend}
\alias{ChooseK.recommend}
\title{Recommend Cluster Number based on Bootstrap Results.}
\usage{
ChooseK.recommend(results, metric, max_opt = FALSE)
}
\arguments{
\item{results}{Data.frame of bootstrap results}

\item{metric}{String, metric of interest.}

\item{max_opt}{Is maximizing the metric optimal?}
}
\value{
Data.frame containing:
\itemize{
  \item `Metric` name.
  \item Optimal cluster number `k_opt` and metric value `Metric_opt`.
  \item 1SE cluster number `k_1se` and metric value `Metric_1se`.
}
}
\description{
Recommend Cluster Number based on Bootstrap Results.
}
