% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{ClustQual.partition_by_clust}
\alias{ClustQual.partition_by_clust}
\title{Partition Data by Cluster Assignment.}
\usage{
ClustQual.partition_by_clust(data, assign)
}
\arguments{
\item{data}{Numeric data.matrix.}

\item{assign}{Cluster assignments.}
}
\value{
List of numeric data.matrices, separated by cluster assignment.
}
\description{
Partition Data by Cluster Assignment.
}
