% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustQual}
\alias{clustQual}
\title{Cluster Quality}
\usage{
clustQual(M)
}
\arguments{
\item{M}{Object of class mix.}
}
\value{
A list containing the cluster quality metrics.
}
\description{
Evaluates cluster quality. Returns the following metrics:
\itemize{
\item BIC: Bayesian Information Criterion, lower value indicates better clustering quality.
\item CHI: Calinski-Harabaz Index, higher value indicates better clustering quality.
\item DBI: Davies-Bouldin, lower value indicates better clustering quality.
\item SIL: Silhouette Width, higher value indicates better clustering quality.
}
}
\examples{
\dontrun{
set.seed(100);
# Data generation
Y = rGMM(n=500,d=3,k=4);
M = fit.GMM(Y=Y,k=3);
# Clustering quality
Q = clustQual(M);
}
}
\seealso{
See \code{\link{chooseK}} for using quality metrics to choose the cluster number.
}
