% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_Imputation.R
\name{CombineMIs}
\alias{CombineMIs}
\title{Combine Multiple Imputations}
\usage{
CombineMIs(points, covs)
}
\arguments{
\item{points}{List of point estimates, potentially vector valued.}

\item{covs}{List of sampling covariances, potentially matrix valued.}
}
\value{
List containing the final point estimate (`point`) and 
  sampling covariance (`cov`).
}
\description{
Combines point estimates and standard errors across multiple imputations.
}
