% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov}
\alias{cov}
\title{Correlation}
\usage{
cov(A, B, cor = FALSE)
}
\arguments{
\item{A}{First matrix.}

\item{B}{Second matrix.}

\item{cor}{Return correlation matrix?}
}
\value{
Numeric matrix.
}
\description{
Calculates the correlation between two vectors.
}
