% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{DavBou.clust_diameter}
\alias{DavBou.clust_diameter}
\title{Mean Cluster Diameter}
\usage{
DavBou.clust_diameter(clust, mean)
}
\arguments{
\item{clust}{Numeric data.matrix.}

\item{mean}{Cluster mean.}
}
\value{
Scalar mean diameter.
}
\description{
Mean Cluster Diameter
}
