% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigSym}
\alias{eigSym}
\title{Eigenvalues of Symmetric Matrix.}
\usage{
eigSym(A)
}
\arguments{
\item{A}{symmetric matrix.}
}
\value{
Numeric vector.
}
\description{
Calculates the eigenvalues of a symmetric matrix.
}
