% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expectations.R
\name{ExpResidOP}
\alias{ExpResidOP}
\title{Expected Residual Outer Product}
\usage{
ExpResidOP(Y1, m1, m0, S0, g = NULL)
}
\arguments{
\item{Y1}{Data for observations with missingness.}

\item{m1}{New mean.}

\item{m0}{Old mean.}

\item{S0}{Old covariance.}

\item{g}{Responsibilities.}
}
\value{
Numeric matrix, the responsibility-weighted, cumulative,
 expected residual outer product.
}
\description{
Calculates the expected residual outer product.
}
