% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastIP}
\alias{fastIP}
\title{Matrix Inner Product}
\usage{
fastIP(A, B)
}
\arguments{
\item{A}{Numeric matrix.}

\item{B}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the product \eqn{A'B}.
}
