% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastOP}
\alias{fastOP}
\title{Fast Outer Product}
\usage{
fastOP(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Numeric matrix.
}
\description{
Calculates the outer product \eqn{xy'}.
}
