% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastQF}
\alias{fastQF}
\title{Quadratic Form}
\usage{
fastQF(X, A)
}
\arguments{
\item{X}{Numeric matrix.}

\item{A}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the quadratic form \eqn{X'AX}.
}
