% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{fit.mix.miss.impute}
\alias{fit.mix.miss.impute}
\title{Imputation for Mixutre of MVNs with Missingness.}
\usage{
fit.mix.miss.impute(split_data, theta)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{theta}{List containing the current `means`, `covs`, `pi`, and `gamma`.}
}
\value{
Data.matrix, in the same order as the original data, with missing values
  imputed to their expectations.
}
\description{
Imputation for Mixutre of MVNs with Missingness.
}
