% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{fit.mix.miss.init}
\alias{fit.mix.miss.init}
\title{Parameter Initialization for Mixture of Multivariate Normals.}
\usage{
fit.mix.miss.init(split_data, k, init_means, init_covs, init_props)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{k}{Number of mixture components.}

\item{init_means}{Optional list of initial mean vectors.}

\item{init_covs}{Optional list of initial covariance matrices.}

\item{init_props}{Optional vector of initial cluster proportions.}
}
\description{
Parameter Initialization for Mixture of Multivariate Normals.
}
