% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{fit.mix.miss.update.means}
\alias{fit.mix.miss.update.means}
\title{Mean Update for Mixture of MVNs with Missingness.}
\usage{
fit.mix.miss.update.means(split_data, means, covs, gamma)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{means}{List of component means.}

\item{covs}{List of component covariances.}

\item{gamma}{List of component responsibilities.}
}
\value{
List containing the updated component means.
}
\description{
Mean Update for Mixture of MVNs with Missingness.
}
