% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{fit.mix.miss.update}
\alias{fit.mix.miss.update}
\title{Parameter Update for Mixutre of MVNs with Missingness.}
\usage{
fit.mix.miss.update(split_data, theta, fix_means)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{theta}{List containing the current `means`, `covs`, `pi`, and `gamma`.}

\item{fix_means}{Fix the mean to its starting value? Must initialize.}
}
\value{
List containing:
\itemize{
  \item The updated `mean`, `cov`, `pi`, and `gamma`.
  \item The initial `old_obj` and final `new_obj` EM objective. 
  \item The increase in the EM objective `delta`. 
}
}
\description{
Parameter Update for Mixutre of MVNs with Missingness.
}
