% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{fit.mix}
\alias{fit.mix}
\title{Fit Multivariate Mixture Distribution}
\usage{
fit.mix(Y, k = 2, M0 = NULL, S0 = NULL, pi0 = NULL, maxit = 100,
  eps = 1e-06, report = F, parallel = F)
}
\arguments{
\item{Y}{Numeric data matrix.}

\item{k}{Number of mixture components. Defaults to 2.}

\item{M0}{Optional list of initial mean vectors.}

\item{S0}{Optional list of initial covariance matrices.}

\item{pi0}{Optional vector of initial cluster proportions.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report fitting progress?}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
Object of class \code{mix} containing the estimated
}
\description{
Given a matrix of random vectors, estimates the parameters for a mixture of
multivariate normal distributions. Accommodates arbitrary patterns of 
missingness, provided the elements are missing at random (MAR).
}
