% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{fit.mvn.miss.init}
\alias{fit.mvn.miss.init}
\title{Parameter Initialization for Single Component Multivariate Normal with Missingness.}
\usage{
fit.mvn.miss.init(data_comp, init_mean, init_cov)
}
\arguments{
\item{data_comp}{Complete cases.}

\item{init_mean}{Optional initial mean vector.}

\item{init_cov}{Optional initial covariance matrix.}
}
\value{
List containing the initialized `mean` and `cov`.
}
\description{
Parameter Initialization for Single Component Multivariate Normal with Missingness.
}
