% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{fit.mvn.miss}
\alias{fit.mvn.miss}
\title{Estimation for a Single Component Multivariate Normal with Missingness.}
\usage{
fit.mvn.miss(data, init_mean, fix_mean, init_cov, maxit, eps, report)
}
\arguments{
\item{data}{Numeric data matrix.}

\item{init_mean}{Optional initial mean vector.}

\item{fix_mean}{Fix the means to their starting value? Must initialize.}

\item{init_cov}{Optional initial covariance matrix.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report fitting progress?}
}
\value{
List containing the final `Mean`, `Covariance`, EM `Objective` and a
  `Completed` data matrix.
}
\description{
Estimation for a Single Component Multivariate Normal with Missingness.
}
