% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{fit.mvn.miss.update}
\alias{fit.mvn.miss.update}
\title{Parameter Update for Single Component Multivariate Normal with Missingness.}
\usage{
fit.mvn.miss.update(split_data, theta, fix_mean)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{theta}{List containing the current `mean` and `cov`.}

\item{fix_mean}{Fix the mean to its starting value? Must initialize.}
}
\value{
List containing:
\itemize{
  \item The updated `mean` and `cov`.
  \item The initial `old_obj` and final `new_obj` EM objective. 
  \item The increase in the EM objective `delta`. 
}
}
\description{
Parameter Update for Single Component Multivariate Normal with Missingness.
}
