% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{fit.mvn.no_miss}
\alias{fit.mvn.no_miss}
\title{Estimation for a Single Component Multivariate Normal with No Missingness.}
\usage{
fit.mvn.no_miss(data, init_mean, fix_mean)
}
\arguments{
\item{data}{Numeric data matrix.}

\item{init_mean}{Optional initial mean vector.}

\item{fix_mean}{Fix the means to their starting value? Must initialize}
}
\value{
List containing the new `Mean`, `Covariance`, and log-likelihood
  `Objective`.
}
\description{
Estimation for a Single Component Multivariate Normal with No Missingness.
}
