% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mvn.R
\name{fit.mvn}
\alias{fit.mvn}
\title{Fit Multivariate Normal Distribution}
\usage{
fit.mvn(Y, m0 = NULL, S0 = NULL, maxit = 100, eps = 1e-06, report = T)
}
\arguments{
\item{Y}{Numeric data matrix.}

\item{m0}{Optional initial mean vector.}

\item{S0}{Optional initial covariance matrix.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report fitting progress?}
}
\value{
List containing the estimated mean and covariance. If missingness is 
  present, the final value of the EM objective is also provided.
}
\description{
Given a matrix of random vectors, estimates the parameters of a multivariate 
normal distribution Accommodates arbitrary patterns of missingness, provided
the elements are missing at random (MAR).
}
