% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_Imputation.R
\name{GenImputation}
\alias{GenImputation}
\title{Generate Imputation}
\usage{
GenImputation(fit)
}
\arguments{
\item{fit}{Fitted model.}
}
\value{
Numeric matrix with missing values imputed.
}
\description{
Generates a stochastic imputation of a data set from a fitted data set.
}
