% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{imputeOP}
\alias{imputeOP}
\title{Imputes Sum of Residual Outer Products}
\usage{
imputeOP(Y1, g, m1, m0, S0)
}
\arguments{
\item{Y1}{Data for observations with missingness.}

\item{g}{Responsibilities.}

\item{m1}{New mean.}

\item{m0}{Old mean.}

\item{S0}{Old covariance.}
}
\value{
Numeric matrix, the sum of the individual conditional expectations.
}
\description{
Calculates the expected sum of residual outer products,
conditional on the observed elements.
}
