% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{imputeSum}
\alias{imputeSum}
\title{Impute Sum of Observations with Missingness}
\usage{
imputeSum(Y1, g, m0, S0)
}
\arguments{
\item{Y1}{Incomplete observations.}

\item{g}{Responsibilities.}

\item{m0}{Previous mean.}

\item{S0}{Previous covariance.}
}
\value{
Numeric vector, the sum of the individual conditional expectations.
}
\description{
Calculates the expected sum of response vectors, conditional on the
observed elements.
}
