% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matCov}
\alias{matCov}
\title{Covariance}
\usage{
matCov(A, B, corMat = FALSE)
}
\arguments{
\item{A}{NxP matrix.}

\item{B}{NxQ matrix.}

\item{corMat}{Return correlation matrix? If false, returns a covariance matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the correlation between two matrices.
}
