% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matOP}
\alias{matOP}
\title{Fast Outer Product}
\usage{
matOP(X, Y)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the outer product \eqn{XY'}.
}
