% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{Maximization}
\alias{Maximization}
\title{Maximization Iteration.}
\usage{
Maximization(theta0, Update, maxit, eps, report)
}
\arguments{
\item{theta0}{List containing the initial values of `mean` and `cov`.}

\item{Update}{Function to iteratively update theta0.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report fitting progress?}
}
\value{
List containing the updated `mean` and `cov`, the final EM objective
  `new_obj`.
}
\description{
Maximization Iteration.
}
