% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{mean.mvn}
\alias{mean.mvn}
\title{Mean for Fitted MVN Model}
\usage{
\method{mean}{mvn}(x, ...)
}
\arguments{
\item{x}{A \code{mvn} object.}

\item{...}{Unused.}
}
\description{
Mean for Fitted MVN Model
}
