% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGMM-package.R
\docType{package}
\name{MGMM-package}
\alias{MGMM-package}
\alias{MGMM}
\alias{_PACKAGE}
\title{Missingness-Aware Gaussian Mixture Models}
\description{
Estimate and classify using Gaussian mixture models (GMMs) when the input
contains missing values. Uses an expectation–conditional maximization (ECM)
algorithm with support for full component covariances and optional ridge
regularization. See McCaw et al. (2022) <doi:10.1186/s12859-022-04740-9>.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
