% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{mix-class}
\alias{mix-class}
\title{Mixture Model Class}
\description{
Defines the class returned by functions that 
fit mixture models.
}
\section{Slots}{

\describe{
\item{\code{Components}}{Components.}

\item{\code{Means}}{Fitted cluster means.}

\item{\code{Covariances}}{Fitted cluster covariances.}

\item{\code{Proportions}}{Fitted cluster proportions.}

\item{\code{Objective}}{Final value of the EM objective.}

\item{\code{Responsibilities}}{Final cluster responsibilities.}

\item{\code{Assignments}}{Final cluster assignments.}
}}

