% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{show,mix-method}
\alias{show,mix-method}
\title{Show for Fitted Mixture Models}
\usage{
\S4method{show}{mix}(object)
}
\arguments{
\item{object}{A \code{mix} object.}
}
\description{
Show for Fitted Mixture Models
}
