% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{MixClusterAssign}
\alias{MixClusterAssign}
\title{Cluster Assignment for Mixutre of MVNs with Missingness.}
\usage{
MixClusterAssign(split_data, theta)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{theta}{List containing the current `means`, `covs`, `pi`, and `gamma`.}
}
\value{
List containing:
\itemize{
  \item Matrix of cluster `Assignments`.
  \item Matrix of `Density` evaluations. 
  \item Matrix of cluster `Responsibilities`.
}
}
\description{
Cluster Assignment for Mixutre of MVNs with Missingness.
}
