% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{MixClusterSizes}
\alias{MixClusterSizes}
\title{Cluster Sizes for a Mixutre of MVNs.}
\usage{
MixClusterSizes(split_data, gamma)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{gamma}{List cof component responsibilities.}
}
\description{
Cluster Sizes for a Mixutre of MVNs.
}
