% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{MixEMObj}
\alias{MixEMObj}
\title{EM Objective for a Mixture of MVNs.}
\usage{
MixEMObj(cluster_sizes, pi, covs, resid_ops)
}
\arguments{
\item{cluster_sizes}{Cluster sizes.}

\item{pi}{Cluster proportions}

\item{covs}{List of component covariances.}

\item{resid_ops}{List of residual outer products.}
}
\description{
EM Objective for a Mixture of MVNs.
}
