% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Mix_MVN.R
\name{MixResidOP}
\alias{MixResidOP}
\title{Expected Residual Outer Product for a Mixutre of MVNs.}
\usage{
MixResidOP(split_data, new_means, old_means, covs, gamma)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{new_means}{List of updated means.}

\item{old_means}{List of previous means.}

\item{covs}{List of component covariances.}

\item{gamma}{List cof component responsibilities.}
}
\value{
List of k expected residual outer products.
}
\description{
Expected Residual Outer Product for a Mixutre of MVNs.
}
