% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Mix_MVN.R
\name{MixUpdateMeans}
\alias{MixUpdateMeans}
\title{Mean Update for Mixture of MVNs with Missingness.}
\usage{
MixUpdateMeans(split_data, means, covs, gamma)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{means}{List of component means.}

\item{covs}{List of component covariances.}

\item{gamma}{List of component responsibilities.}
}
\value{
List containing the updated component means.
}
\description{
Paper eq. (7): \eqn{\mu_j^{(r+1)} = (1/n_j) \sum_i \gamma_{ij} \hat{y}_{ij}}{mu_j^(r+1) = (1/n_j) sum_i gamma_ij y_ij} (responsibility-weighted average of working responses).
}
