% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MMP}
\alias{MMP}
\title{Matrix Matrix Product}
\usage{
MMP(A, B)
}
\arguments{
\item{A}{Numeric matrix.}

\item{B}{Numeric matrix.}
}
\value{
Numeric matrix.
}
\description{
Calculates the product \eqn{AB}.
}
