% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/About.R
\docType{package}
\name{MNMix}
\alias{MNMix}
\alias{MNMix-package}
\title{MNMix: Multivariate Normal Mixture Models}
\description{
Parameter estimation for multivariate normal mixture models with one or more
components, and arbitrary patterns of missingness, provided the data are
missing at random.
}
\author{
Zachary R. McCaw
}
