% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\docType{class}
\name{mvn-class}
\alias{mvn-class}
\title{Multivariate Normal Model Class}
\description{
Defines a class to hold multivariate normal models.
}
\section{Slots}{

\describe{
\item{\code{Completed}}{Completed data, with missing values imputed to their
posterior expectations.}

\item{\code{Covariance}}{Fitted covariance matrix.}

\item{\code{Data}}{Original data, with missing values present.}

\item{\code{Mean}}{Fitted mean vector.}

\item{\code{Objective}}{Final value of the EM objective.}
}}

