% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{show,mvn-method}
\alias{show,mvn-method}
\title{Show for Multivariate Normal Models}
\usage{
\S4method{show}{mvn}(object)
}
\arguments{
\item{object}{A \code{mvn} object.}
}
\description{
Show for Multivariate Normal Models
}
