% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Single_MVN.R
\name{PartitionData}
\alias{PartitionData}
\title{Paratition a Data.frame by Missingness cases.}
\usage{
PartitionData(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
List containing:
\itemize{
  \item The original row and column names: `orig_row_names`, `orig_col_names`.
  \item The original row and column numbers: `n_row` and `n_col`.
  \item The complete cases `data_comp`, and their indices `idx_comp` in `data`.
  \item The incomplete cases `data_incomp`, and their indices `idx_incomp` in `data`.
  \item The empty cases `data_empty`, and their indices `idx_empty` in `data`.
  \item Counts of complete `n0`, incomplete `n1`, and empty `n2` cases. 
}
}
\description{
Paratition a Data.frame by Missingness cases.
}
