% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{print.mvn}
\alias{print.mvn}
\title{Print Fitted MVN Model}
\usage{
\method{print}{mvn}(x, ...)
}
\arguments{
\item{x}{A \code{mvn} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Fitted MVN Model
}
