% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{Q.mix}
\alias{Q.mix}
\title{EM Objective Function}
\usage{
Q.mix(M, S, pi, Y0 = NULL, Y1 = NULL, G0, G1, V0s, V1s)
}
\arguments{
\item{M}{Current means. Unnecessary if outer products are provided.}

\item{S}{Current covariance matrices.}

\item{pi}{Current cluster proportions.}

\item{Y0}{Complete observations.}

\item{Y1}{Incomplete observations.}

\item{G0}{Responsibilities for complete observations.}

\item{G1}{Responsibilities for incomplete observations.}

\item{V0s}{Outer products for complete observations.}

\item{V1s}{Outer products for incomplete observations.}
}
\value{
Scalar value of the objective.
}
\description{
EM Objective Function
}
