% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mvn.R
\name{Q.mvn}
\alias{Q.mvn}
\title{EM Objective Function}
\usage{
Q.mvn(m, S, Y0 = NULL, Y1 = NULL, V0, V1)
}
\arguments{
\item{m}{Current mean, unnecessary if outer products are provided.}

\item{S}{Current covariance matrix.}

\item{Y0}{Complete observations.}

\item{Y1}{Incomplete observations.}

\item{V0}{Outer products for complete observations.}

\item{V1}{Outer products for incomplete observations.}
}
\value{
Scalar value of the objective.
}
\description{
Expectation of the complete data log likelihood, given the
observed data, for a multivariate normal distribution with
a single component.
}
