% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Expectations.R
\name{Responsibility.eval_dens_incomp}
\alias{Responsibility.eval_dens_incomp}
\title{Evaluate the Density of an Incomplete Observations}
\usage{
Responsibility.eval_dens_incomp(y, means, covs, pi)
}
\arguments{
\item{y}{Vector with missing elements.}

\item{means}{List of mean vectors.}

\item{covs}{List of covariances matrices.}

\item{pi}{Vector of cluster proportions.}
}
\value{
Numeric density of observed elements.
}
\description{
Evaluate the Density of an Incomplete Observations
}
