% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expectations.R
\name{Responsibility}
\alias{Responsibility}
\title{Responsibilities}
\usage{
Responsibility(M, S, pi, Y0 = NULL, Y1 = NULL, parallel = F)
}
\arguments{
\item{M}{List of mean vectors.}

\item{S}{List of covariances matrices.}

\item{pi}{Vector of cluster proportions.}

\item{Y0}{Matrix of complete observations.}

\item{Y1}{Matrix of incomplete observations.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
List of numeric matrices, \code{G0} contains the responsibilities for
  complete observations, and \code{G1} for incomplete observations. Each 
  responsibility matrix is formatted with observations as rows, and mixture 
  components as columns.
}
\description{
Calculates the posterior probability of cluster membership given the observed
data.
}
