% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rGMM}
\alias{rGMM}
\title{Data Generation from Multivariate Normal Mixture Models}
\usage{
rGMM(n, d = 2, k = 1, pi, m = 0, M, S)
}
\arguments{
\item{n}{Observations.}

\item{d}{Observation dimension.}

\item{k}{Number of mixture components. Defaults to 1.}

\item{pi}{Mixture proportions. If omitted, components are assumed 
equi-probable.}

\item{m}{Proportion of elements missing, \eqn{m\in[0,1)}.}

\item{M}{Either a prototype mean vector, or a list of mean vectors. Defaults
to the zero vector.}

\item{S}{Either a prototype covariance matrix, or a list of covariance 
matrices. Defaults to the identity matrix.}
}
\value{
Numeric matrix with observations as rows. Row numbers specify the
  true cluster assignments.
}
\description{
Generates an \eqn{n\times d} matrix of multivariate normal random vectors
with observations as rows. If \eqn{k=1}, all observations belong to the same
cluster. If \eqn{k>1} the observations are generated via two-step precoedure.
First, the cluster membership is drawn from a multinomial distribution, with
mixture proportions specified by \code{pi}. Conditional on cluster
membership, the observation is drawn from a multivariate normal distribution,
with cluster-specific mean and covariance. The cluster means are provided
using \code{M}, and the cluster covariance matrices are provided using
\code{S}. If \eqn{m>0}, missingness is introduced, completely at random, by
setting that proportion of elements in the data matrix to \code{NA}.
}
\examples{
\dontrun{
set.seed(100);
# Single component without missingness
# Bivariate normal observations
Sigma = matrix(c(1,0.5,0.5,1),nrow=2);
Y = rGMM(n=1e3,d=2,k=1,M=c(2,2),S=Sigma);

# Single component with missingness
# Trivariate normal observations
M = list(c(-2,-2,-2),c(2,2,2));
Sigma = matrix(c(1,0.5,0.5,0.5,1,0.5,0.5,0.5,1),nrow=3);
Y = rGMM(n=1e3,d=3,k=2,M=M,S=Sigma);

# Two components without missingness
# Trivariate normal observations
Means = list(c(-2,-2,-2),c(2,2,2));
Sigma = matrix(c(1,0.5,0.5,0.5,1,0.5,0.5,0.5,1),nrow=3);
Y = rGMM(n=1e3,d=3,k=2,M=Means,S=Sigma);

# Four components with missingness
# Bivariate normal observations
M = list(c(2,2),c(2,-2),c(-2,2),c(-2,-2));
S = 0.5*diag(2);
Y = rGMM(n=1000,d=2,k=4,pi=c(0.35,0.15,0.15,0.35),m=0.1,M=M,S=S);
}
}
\seealso{
For estimation, see \code{\link{fit.GMM}}.
}
