% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SchurC}
\alias{SchurC}
\title{Schur complement}
\usage{
SchurC(I11, I22, I12)
}
\arguments{
\item{I11}{Information of target parameter}

\item{I22}{Information of nuisance parameter}

\item{I12}{Cross information between target and nuisance parameters}
}
\value{
Numeric matrix.
}
\description{
Calculates the efficient information \eqn{I_{11}-I_{12}I_{22}^{-1}I_{21}}.
}
