% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expectations.R
\name{SumWorkResp}
\alias{SumWorkResp}
\title{Sum of Working Response Vectors}
\usage{
SumWorkResp(Y1, m0, S0, g = NULL)
}
\arguments{
\item{Y1}{Incomplete observations.}

\item{m0}{Previous mean.}

\item{S0}{Previous covariance.}

\item{g}{Responsibilities}
}
\value{
Numeric vector, the responsibility-weighted cumulative working
  response vector.
}
\description{
Calculate the sum of the working response vectors.
}
