% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{Update.mix}
\alias{Update.mix}
\title{Update Function}
\usage{
Update.mix(theta, Y0 = NULL, Y1 = NULL, parallel = F)
}
\arguments{
\item{theta}{Current parameter list.}

\item{Y0}{Complete observations.}

\item{Y1}{Incomplete observations.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
List containing the proposed means, covariances, cluster proportions,
  responsibilities, and objective function.
}
\description{
EM parameter update for the normal mixture model with missingness.
}
