% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mvn.R
\name{Update.mvn}
\alias{Update.mvn}
\title{Update Function}
\usage{
Update.mvn(theta, Y0 = NULL, Y1 = NULL)
}
\arguments{
\item{theta}{Current parameter list, containing the mean, covariance, and
objective.}

\item{Y0}{Complete observations.}

\item{Y1}{Incomplete observations.}
}
\value{
List containing the proposed mean, covariance, and objective.
}
\description{
EM parameter update for a multivariate normal distribution
with a single component.
}
