% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{vcov.mix}
\alias{vcov.mix}
\title{Covariance for Fitted GMM}
\usage{
\method{vcov}{mix}(object, ...)
}
\arguments{
\item{object}{A \code{mix} object.}

\item{...}{Unused.}
}
\description{
Covariance for Fitted GMM
}
