% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{vcov.mvn}
\alias{vcov.mvn}
\title{Covariance for Fitted MVN Model}
\usage{
\method{vcov}{mvn}(object, ...)
}
\arguments{
\item{object}{A \code{mvn} object.}

\item{...}{Unused.}
}
\description{
Covariance for Fitted MVN Model
}
