% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_Clustering.R
\name{CalHar}
\alias{CalHar}
\title{Calinski-Harabasz Index}
\usage{
CalHar(data, assign, means)
}
\arguments{
\item{data}{Numeric matrix of observations.}

\item{assign}{Vector of cluster assignments.}

\item{means}{List of cluster mean vectors.}
}
\value{
Numeric scalar; higher values indicate better separation.
}
\description{
Calculates the Calinski-Harabasz index (ratio of between-cluster to
within-cluster dispersion; higher values indicate better separation).
}
