% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{CalHar}
\alias{CalHar}
\title{Calinski-Harabaz Index}
\usage{
CalHar(Y, a, m)
}
\arguments{
\item{Y}{Observations}

\item{a}{Assignments}

\item{m}{List of cluster means}
}
\value{
Scalar.
}
\description{
Calculates the Calinski-Harabaz index.
}
