% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_Clustering.R
\name{CalHar}
\alias{CalHar}
\title{Calinski-Harabaz Index}
\usage{
CalHar(data, assign, means)
}
\arguments{
\item{data}{Observations.}

\item{assign}{Assignments.}

\item{means}{List of cluster means.}
}
\value{
Scalar metric.
}
\description{
Calculates the Calinski-Harabaz index.
}
