% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{DavBou}
\alias{DavBou}
\title{Davies-Bouldin Index}
\usage{
DavBou(Y, a, M)
}
\arguments{
\item{Y}{Observations}

\item{a}{Assignments}

\item{M}{List of cluster means}
}
\value{
Scalar.
}
\description{
Calculates the Davies-Bouldin index.
}
