% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_Clustering.R
\name{DavBou}
\alias{DavBou}
\title{Davies-Bouldin Index}
\usage{
DavBou(data, assign, means)
}
\arguments{
\item{data}{Observations}

\item{assign}{Assignments}

\item{means}{List of cluster means}
}
\value{
Scalar index.
}
\description{
Calculates the Davies-Bouldin index.
}
