% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Expectations.R
\name{ExpResidOP}
\alias{ExpResidOP}
\title{Expected Residual Outer Product}
\usage{
ExpResidOP(data_incomp, new_mean, old_mean, old_cov, gamma = NULL)
}
\arguments{
\item{data_incomp}{Data for observations with missingness.}

\item{new_mean}{New mean.}

\item{old_mean}{Initial mean.}

\item{old_cov}{Initial covariance.}

\item{gamma}{Responsibilities.}
}
\value{
Numeric matrix, the responsibility-weighted, cumulative,
 expected residual outer product.
}
\description{
Calculates the expected residual outer product.
}
