% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVN.R
\name{fit.mvn}
\alias{fit.mvn}
\title{Fit Multivariate Normal Distribution}
\usage{
fit.mvn(Y, m0 = NULL, fix.means = F, S0 = NULL, maxit = 100,
  eps = 1e-06, report = T)
}
\arguments{
\item{Y}{Numeric data matrix.}

\item{m0}{Optional initial mean vector.}

\item{fix.means}{Fix the means to their starting value? Must provide initial
values.}

\item{S0}{Optional initial covariance matrix.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Minimum acceptable increment in the EM objective.}

\item{report}{Report fitting progress?}
}
\value{
List containing the estimated mean, covariance, and objective.
}
\description{
Given a matrix of random vectors, estimates the mean and covariance of a
multivariate normal distribution. Accommodates arbitrary patterns of
missingness at random.
}
