% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{logLik.mix}
\alias{logLik.mix}
\title{Log likelihood for Fitted GMM}
\usage{
\method{logLik}{mix}(object, ...)
}
\arguments{
\item{object}{A \code{mix} object.}

\item{...}{Unused.}
}
\description{
Log likelihood for Fitted GMM
}
