% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{logLik.mvn}
\alias{logLik.mvn}
\title{Log likelihood for Fitted MVN Model}
\usage{
\method{logLik}{mvn}(object, ...)
}
\arguments{
\item{object}{A \code{mvn} object.}

\item{...}{Unused.}
}
\description{
Log likelihood for Fitted MVN Model
}
