% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_About.R
\docType{package}
\name{MGMM}
\alias{MGMM}
\title{MGMM: Missingness Aware Gaussian Mixture Models}
\description{
Parameter estimation and classification via Gaussian Mixture Models, allowing
for missingness in the input vectors. See \code{\link{fit.GMM}} for estimating
the GMM, and \code{\link{ChooseK}} for selecting the number of clusters. See 
\code{\link{rGMM}} for simulating data from a GMM.
}
\author{
Zachary R. McCaw
}
