% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\docType{class}
\name{mix-class}
\alias{mix-class}
\title{Mixture Model Class}
\description{
Defines the class returned by the fitting functions.
}
\section{Slots}{

\describe{
\item{\code{Assignments}}{Maximum a posteriori assignment.}

\item{\code{Completed}}{Completed data, with missing values replaced by posterior expectations.}

\item{\code{Components}}{Components.}

\item{\code{Covariances}}{Fitted cluster covariances.}

\item{\code{Density}}{Component density at observation.}

\item{\code{Means}}{Fitted cluster means.}

\item{\code{Objective}}{Final value of the EM objective.}

\item{\code{Proportions}}{Fitted cluster proportions.}

\item{\code{Responsibilities}}{Posterior membership probabilities.}
}}

