% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/About.R
\docType{package}
\name{MNMix}
\alias{MNMix}
\alias{MNMix-package}
\title{MGMM: Missingness Aware Gaussian Mixture Models}
\description{
Parameter estimation and classification via Gaussian Mixture Models, allowing
for missingness in the input vectors. For estimation of the GMM, see
\code{\link{fit.GMM}}. For data generation from a GMM, see
\code{\link{rGMM}}.
}
\author{
Zachary R. McCaw
}
