% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/About.R
\docType{package}
\name{MNMix}
\alias{MNMix}
\alias{MNMix-package}
\title{MNMix: Multivariate Normal Mixture Models}
\description{
Parameter estimation for Gausian Mixture Models (GMMs) with one or more mixutre
components, and arbitrary patterns of missingness at random in the input
vectors. For estimation of the GMM, see \code{\link{fit.MNMix}}. For data generation
from a GMM, see \code{\link{rMNMix}}.
}
\author{
Zachary R. McCaw
}
