% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Class.R
\name{print.mix}
\alias{print.mix}
\title{Print for Fitted Mixture Model}
\usage{
\method{print}{mix}(x, ...)
}
\arguments{
\item{x}{A \code{mix} object.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{mix}.
}
