% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Partition_Data.R
\name{ReconstituteData}
\alias{ReconstituteData}
\title{Reconstitute Data}
\usage{
ReconstituteData(split_data)
}
\arguments{
\item{split_data}{Split data are returned by \code{\link{PartitionData}}.}
}
\value{
Numeric matrix.
}
\description{
Reassembles a data matrix split by missingness pattern.
}
