% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Expectations.R
\name{Responsibility}
\alias{Responsibility}
\title{Responsibilities}
\usage{
Responsibility(split_data, means, covs, pi)
}
\arguments{
\item{split_data}{Data partitioned by missingness.}

\item{means}{List of mean vectors.}

\item{covs}{List of covariances matrices.}

\item{pi}{Vector of cluster proportions.}
}
\value{
List containing:
\itemize{
  \item k Number of mixture components.
  \item Density evaluations `dens_eval0` and responsibilities `gamma0` 
    for complete cases.
  \item Density evaluations `dens_eval1` and responsibilities `gamma1` 
    for incomplete cases. 
}
}
\description{
Calculates the posterior probability of cluster membership given the observed
data.
}
